+++ config/deploy/production.rb
@@ -10,7 +10,7 @@ namespace :deploy do
   namespace :check do
     before :linked_files, :set_master_key do
       on roles(:app), in: :sequence, wait: 10 do
-        upload! "config/settings/production.yml", "#{shared_path}/config/settings/production.yml"
+        upload! "config/settings/production.local.yml", "#{shared_path}/config/settings/production.yml"
       end
     end
   end