+++ db/migrate/20140325150638_devise_create_users.rb
@@ -36,7 +36,7 @@ class DeviseCreateUsers < ActiveRecord::Migration[4.2]
       # t.string   :unlock_token # Only if unlock strategy is :email or :both
       t.datetime :locked_at
 
-      ## CAS Authenticable
+      ## Login field
       t.string :username, null: false, default: '', index: true, unique: true
 
       ## other