+++ db/migrate/20210930173228_add_column_metadata_to_audit.rb
@@ -1,0 +1,8 @@
+class AddColumnMetadataToAudit < ActiveRecord::Migration[6.1]
+  def change
+    change_table :audits, bulk: true do |t|
+      t.column :metadata, :jsonb, null: false, default: {}
+      t.index  :metadata, using: :gin
+    end
+  end
+end